boost & MPPT not ok

+++ ctl/component/digital_power/src/ctl_dp_mppt.c
@@ -24,7 +24,7 @@ void ctl_init_mppt_PnO_algo(
     ctl_init_lp_filter(&mppt->power_filter, freq_ctrl, freq_mppt * 6.0f);
 
     // divider
-    ctl_init_divider(&mppt->divider, freq_ctrl / freq_mppt);
+    ctl_init_divider(&mppt->divider, (uint32_t)(freq_ctrl / freq_mppt));
 
     // initial voltage
     mppt->v_ref = voltage_0;
